move DataStorage to separate file
